requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
astroid==2.4.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.3
Flask==2.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
isort==4.3.21
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.4.3
MarkupSafe==2.1.1
mccabe==0.6.1
numpy==1.22.4
Pillow==9.1.1
# pygame==2.0.1
pylint==2.5.3
requests==2.27.1
six==1.15.0
toml==0.10.1
torch==1.11.0+cpu
torchvision==0.12.0+cpu
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wincertstore==0.2
wrapt==1.12.1
zipp==3.8.0
